+++ libgloss/mn10300/stat.c
@@ -8,5 +8,5 @@ int
 _stat (const char *path, struct stat *st)
 
 {
-  return TRAP0 (SYS_stat, path, st, 0);
+  return TRAP0 (SYS_stat, (unsigned long) path, (unsigned long) st, 0);
 }
